test/test_build.py
import pytest
from tuxmake.toolchain import Toolchain
from tuxmake.build import build
from tuxmake.build import Build
import tuxmake.exceptions


@pytest.fixture
def output_dir(tmp_path):
    out = tmp_path / "output"
    return out


def test_build(linux, home):
    result = build(linux)
    assert "bzImage" in result.artifacts
    assert (home / ".cache/tuxmake/builds/1/bzImage").exists()


def test_build_with_output_dir(linux, output_dir):
    result = build(linux, output_dir=output_dir)
    assert "bzImage" in result.artifacts
    assert (output_dir / "bzImage").exists()
    assert result.output_dir == output_dir


def test_unsupported_target(linux):
    with pytest.raises(tuxmake.exceptions.InvalidTarget):
        build(linux, targets=["unknown-target"])


def test_kconfig_default(linux, mocker):
    check_call = mocker.patch("subprocess.check_call")
    mocker.patch("tuxmake.build.Build.copy_artifacts")
    mocker.patch("tuxmake.build.Build.cleanup")
    build(linux, targets=["config"])
    assert "defconfig" in check_call.call_args_list[0].args[0]


def test_kconfig_named(linux, mocker):
    check_call = mocker.patch("subprocess.check_call")
    mocker.patch("tuxmake.build.Build.copy_artifacts")
    mocker.patch("tuxmake.build.Build.cleanup")
    build(linux, targets=["config"], kconfig=["fooconfig"])
    assert "fooconfig" in check_call.call_args_list[0].args[0]


def test_kconfig_url(linux, mocker, output_dir):
    response = mocker.MagicMock()
    response.getcode.return_value = 200
    response.read.return_value = b"CONFIG_FOO=y\nCONFIG_BAR=y\n"
    mocker.patch("tuxmake.build.urlopen", return_value=response)

    build(
        linux,
        targets=["config"],
        kconfig=["defconfig", "https://example.com/config.txt"],
        output_dir=output_dir,
    )
    config = output_dir / "config"
    assert "CONFIG_FOO=y\nCONFIG_BAR=y\n" in config.read_text()


def test_kconfig_localfile(linux, tmp_path, output_dir):
    extra_config = tmp_path / "extra_config"
    extra_config.write_text("CONFIG_XYZ=y\nCONFIG_ABC=m\n")
    build(
        linux,
        targets=["config"],
        kconfig=["defconfig", str(extra_config)],
        output_dir=output_dir,
    )
    config = output_dir / "config"
    assert "CONFIG_XYZ=y\nCONFIG_ABC=m\n" in config.read_text()


def test_output_dir(linux, output_dir):
    build(linux, output_dir=output_dir)
    assert [str(f.name) for f in output_dir.glob("*")] == ["config", "bzImage"]


class TestArchitecture:
    def test_x86_64(self, linux):
        result = build(linux, target_arch="x86_64")
        assert "bzImage" in [str(f.name) for f in result.output_dir.glob("*")]

    def test_arm64(self, linux):
        result = build(linux, target_arch="arm64")
        assert "Image.gz" in [str(f.name) for f in result.output_dir.glob("*")]


@pytest.fixture
def builder(linux, output_dir, mocker):
    mocker.patch("tuxmake.build.Build.cleanup")
    mocker.patch("tuxmake.build.Build.copy_artifacts")
    b = Build(linux, output_dir / "tmp", output_dir)
    return b


class TestToolchain:
    # Test that the righ CC= argument is passed. Ideally we want more black box
    # tests that check the results of the build, but for that we need a
    # mechanism to check which toolchain was used to build a given binary (and
    # for test/fakelinux/ to produce real binaries)
    def test_gcc_10(self, builder, mocker):
        check_call = mocker.patch("subprocess.check_call")
        builder.toolchain = Toolchain("gcc-10")
        builder.build("config")
        cmdline = check_call.call_args.args[0]
        cross = builder.arch.makevars["CROSS_COMPILE"]
        assert f"CC={cross}gcc-10" in cmdline

    def test_clang(self, builder, mocker):
        check_call = mocker.patch("subprocess.check_call")
        builder.toolchain = Toolchain("clang")
        builder.build("config")
        cmdline = check_call.call_args.args[0]
        assert "CC=clang" in cmdline
